# docs/api/openapi.yaml
openapi: '3.0.3'
info:
  title: Kalki-Mirror API
  version: 0.2.1
  description: >-
    REST API for the KALKI Tantrik Intelligence Platform.
    Pattern recognition, transit geometry, AI-powered Vedic content,
    and membership management.
  contact:
    name: KALKI Platform
    url: https://www.astrokalki.com
  license:
    name: Proprietary
servers:
  - url: https://www.astrokalki.com
    description: Production
...
paths:
...
  /api/ai/japa-guide:
    post:
      operationId: aiJapaGuide
      tags: [AI]
      summary: AI japa (mantra repetition) guide
      description: >-
        Generates a personalized japa guide for a given mantra.
        No auth. Rate limited 5/min/IP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mantra]
              properties:
                mantra:
                  type: string
                  minLength: 2
                count:
                  type: integer
                  minimum: 1
                  default: 108
                experience:
                  type: string
                  maxLength: 1000
      responses:
        '200':
          description: Japa guide generated
          content:
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /api/ai/explain:
    post:
      operationId: aiExplain
      tags: [AI]
      summary: AI content explanation
      description: >-
        Explains Vedic/tantrik content in beginner or technical style.
        No auth. Rate limited 5/min/IP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  minLength: 20
                  maxLength: 10000
                style:
                  type: string
                  enum: [beginner, technical]
                  default: beginner
      responses:
        '200':
          description: Explanation generated
          content:
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /api/ai/pattern-explain:
    post:
      operationId: aiPatternExplain
      tags: [AI]
      summary: AI pattern explanation
      description: >-
        Explains a specific pattern by its slug with optional context.
        No auth. Rate limited 5/min/IP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [patternSlug]
              properties:
                patternSlug:
                  type: string
                  minLength: 1
                context:
                  type: string
                  maxLength: 500
      responses:
        '200':
          description: Pattern explanation generated
          content:
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /api/ai/draft:
    post:
      operationId: aiDraft
      tags: [AI]
      summary: AI content draft generation
      description: >-
        Generates a content draft for practice, archetype, pattern, research, or codex.
        **Requires authentication.** Rate limited 5/min/IP.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftRequest'
      responses:
        '200':
          description: Draft generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DraftResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

  # ── User ─────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        NextAuth JWT via Bearer token (cookie-based).
        Used for standard authenticated routes.
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        NextAuth JWT with ADMIN, SUPERADMIN, EDITOR, or REVIEWER role required.

  # ── Reusable Responses ──────────────────────────────────
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized — missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ServiceUnavailable:
      description: LLM not configured or returned invalid JSON
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # ── Schemas ─────────────────────────────────────────────
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # Health
...
    DraftType:
      type: string
      enum: [practice, archetype, pattern, research, codex]
    CautionLevel:
      type: string
      enum: [OPEN, MODERATE, HIGH, SEALED]
    Tier:
      type: string
      enum: [prithvi, jal, agni, akash]
    DraftRequest:
      type: object
      required: [type, title]
      properties:
        type:
          $ref: '#/components/schemas/DraftType'
        title:
          type: string
          minLength: 3
          maxLength: 300
        context:
          type: string
          maxLength: 2000
    DraftResponse:
      type: object
      required: [draft, suggestedCaution, suggestedTier, relatedSiddhis]
      properties:
        draft:
          type: string
        suggestedCaution:
          $ref: '#/components/schemas/CautionLevel'
        suggestedTier:
          $ref: '#/components/schemas/Tier'
        relatedSiddhis:
          type: array
          maxItems: 5
          items:
            type: string

    # User
...